sqlite: return an error instead of crashing w/corrupt database
We believe that this is causing a recurring problem in ChromeOS where
the NSS database is corrupted, causing a crash loop in Chrome.

This also replaces an earlier (unused) version of sqlite with the
current one in portage (version 3.6.22).

BUG=chromium:236093
TEST=Built and ran on a device, imported a cert, deleted a cert and
re-imported, each after logging in and out, and rebooting.

Change-Id: I30dd4c01a61f4e1d5c0eeba33144ff07ec3b55e0
Reviewed-on: https://gerrit.chromium.org/gerrit/50904
Tested-by: Greg Spencer <gspencer@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Greg Spencer <gspencer@chromium.org>