Add virtual/chromeos-interface
Introduce a virtual package to facilitate the build-time selection of
lighter-weight, less featured user interfaces in CrOS (e.g. for
embedded devices).  At the same time, introduce a meta-package for the
current, full CrOS UI to allow the packages that implement it to be
included or excluded en masse.

BUG=chromium:242239
TEST=build an image; it should work as it does today
TEST=suite:smoke

Change-Id: If589848d1960c6484a1beda46bab682ad1bf904e
Reviewed-on: https://gerrit.chromium.org/gerrit/56093
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>