Update FreeType 2.4.12
2.4.12 has a new CFF hint handling code contributed from Adobe in
collaboration with Google. We don't use a CFF font on Chrome OS, but
web fonts in CFF can benefit from this change when we begin to honor
native CFF hints in web fonts instead of relying on auto hints.

The new CFF engine is off by default in 2.4.12 (it'll be on in upcoming
2.5.0.) To turn it on, freetype-2.4.12-new-cff-engine-on.patch is applied.
As said above, this does not affect the built-in fonts on Chrome OS, but
turning it on now will help us with CFF web fonts once we change our
hinting method for web fonts in Chrome.

FT_CONFIG_OPTION_OLD_INTERNALS is off by default and there's no need
to disable it again. The sed script is buggy and disabling a feature
(that is off by default) leads to a compilation failure.

Ref: http://google-opensource.blogspot.com/2013/05/got-cff.html

BUG=None
TEST=check if browser/UI text rendering doesn't break

Change-Id: I2b4d6b89b42bb81a8cdf1f95552cb088f3ceba07
Reviewed-on: https://gerrit.chromium.org/gerrit/57381
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Jungshik Shin (jungshik at google) <jshin@chromium.org>
Tested-by: Jungshik Shin (jungshik at google) <jshin@chromium.org>