BlueZ 4: Fix a crash on playstation-peripheral plugin.
If the udev property HID_ID is not set for a device, the playstation
peripheral plugin will crash with a SIGSEGV on a NULL address while
trying to parse the non-existant HID_ID address.

This patch makes the plugin to fail with an error message instead of
crashing.

BUG=chromium:248346
TEST=gmerge bluez

Change-Id: Id978ac61d0b93f893caa41c3915df45b949869d2
Reviewed-on: https://gerrit.chromium.org/gerrit/58132
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>