cros-firmware: Add dependency on a newer version of flashrom
The firmware scripts will soon rely on some new features of flashrom.
Add an explicit dependency to make sure we get the new version.
Specifically, we need this in flashrom:
  ba0827a Implement programmer aliases

Rather than just depend on that version of flashrom, we just add a
dependency on the newest one to get one extra bugfix.

BUG=chrome-os-partner:20261
TEST=gmerge chromeos-base/chromeos-firmware-daisy

Change-Id: I51604aa711682fa74b97a364f1b8003ce5840a3d
Reviewed-on: https://gerrit.chromium.org/gerrit/59070
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>