Prepare the upgrade/addition of pango, harfbuzz & ragel.
Add the pristine copy of ragel and new ebuild files/patches for pango and
harfbuzz with all three masked.

This is a step 1 in upgrading pango (to 1.32.1) and harfbuzz (to 0.9.18).
ragel is added because harfbuzz depends on it during the build time
(no run-time dependency).

We upgrade pango to 1.32.x because 1.30.1 does not use the
latest version of harfbuzz (instead, it uses its built-copy that is
rather old). We can't go to 1.34.x because it also requires upgrading
fontconfig to 2.10.x but we like to avoid upgrading too many pieces
at the same time.

harfbuzz has to be at least 0.9.17 but 0.9.17 ebuild does not allow
us to remove the dependency on icu. So, we go to 0.9.18 to minimize
the local patch. There's no difference between them other than
the build configuration changes.

None of ebuilds works without patches, which will be added in the
next CL. Therefore, all three packages are masked in package.mask.

BUG=chromium:245783
TEST=None (see the above about the needs for local patches).

Change-Id: Id556577f82c9b5cc1034648655ac9606f54d1c4a
Reviewed-on: https://gerrit.chromium.org/gerrit/59083
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Jungshik Shin (jungshik at google) <jshin@chromium.org>
Commit-Queue: Jungshik Shin (jungshik at google) <jshin@chromium.org>