depthcharge: When building a netboot image, use netboot everywhere.
Previously, when building a netboot image, the RO version of depthcharge would
be replaced with the netbooting payload. Because depthcharge was where we
branched to any RW image, that was sufficient to ensure that netboot ran. With
early firmware selection though, the RW version of firmware can run without
any input from depthcharge. This change ensures that those other versions are
also netboot so that what you end up with is always the same.

BUG=chrome-os-partner:20392
TEST=Built and booted regular and netboot images for Link and verified that
they each did what they're supposed to. I wasn't able to boot into an RW image
because of a presumed bug in the EC.
BRANCH=None

Change-Id: I1d3351909ff64828d836effb78787aeaeffd6d2d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/59659
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>