cros-kernel2: Combine build of dtbs with main kernel
Now that dependencies are fixed, we can combine the build of dtbs with
the main kernel buiild.  This saves a bit of build time since we can
avoid the overhead of a 2nd call to kmake and parallelize the
processing of the dtbs.

BUG=chromium:253588
TEST=Time incremental (no-op) build with just this change.  See 2.142
second speedup (44.141 seconds to 41.999 seconds).
TEST=Trybots on kernel and kernel-next pass.
TEST=Build and boot on peach_pit.

Change-Id: I416575de6a83d7c4213266a841397f9deff2ba21
Reviewed-on: https://gerrit.chromium.org/gerrit/59873
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>