Add an option for enabling touch_noise_filter
touch_noise_filter has various parameters tuned specifically for
Pixel Atmel touchscreen and may not work correctly on other touchscreen,
e.g. we found that it incorrectly filters out valid touches on
Acer T232HL because the touch postion value reported on that device
is in a large range (0, 32767) that could trigger far_apart_taps_filter.

So we add a option "EnableTouchNoiseFilter" and plan to only enable it
for Atmel touchscreen for now.

BUG=chromium:255049
TEST=Set the "EnableTouchNoiseFilter" to 0 for Acer T232HL touchscreen
     and the touchscreen works correctly.

Change-Id: Ic905ce583f1a6830c112f25caf88dfdf03f52a29
Reviewed-on: https://gerrit.chromium.org/gerrit/60280
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>