chromeos-bootimage: Revert to using U_BOOT_FDT_USE
This option was introduced to tell chromeos-bootimage which device tree file
to use. Since it is now defined for all platforms, make use of it in the
ebuild.

This seems better than the approach of adding USE flags for all the ARM
boards, as introduced by c23b933. But apart from moving to U_BOOT_FDT_USE,
keep the scheme used by that commit intact.

Also add a new USE flag (USE=build-all-fw) which builds all available images.
This is easier than adding new USE flags for every possible smdk, etc.

BUG=chrome-os-partner:18329
TEST=manual
emerge on various platforms. Will run a trybot on all of them.

Change-Id: I69b39b5d6ea934e5e5b81e753e477f9b5e8ce004
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60625
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>