autotest-chrome ebuild now copies Telemetry source to a writeable temp directory.
A recent change to Telemetry that deleted some files caused chromeOS bots
to fail with "permission denied" errors when trying to delete these files.
This is because the files existed in a non-writable location on disk.

In this change, as part of the src_prepare() step of the autotest-chrome
ebuild, the Telemetry source is copied to a temporary (writable) directory
and then PYTHONPATH is pointed to that location instead.

BUG=chromium:262173
TEST=cbuildbot -G 19790005 -g 62883 --remote x86-generic-full

Change-Id: I2c2733e79d50cb77dae905be10f5bbcb28bb1391
Reviewed-on: https://gerrit.chromium.org/gerrit/62883
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Reviewed-by: Dennis Jeffrey <dennisjeffrey@chromium.org>