laptop-mode-tools: Move USB autosuspend whitelist to a global file
Our current policy is to not autosuspend USB devices unless they have
been explicitly tested and whitelisted to work well with that (only done
for internal devices). This is enforced through the chromeos-bsp-*
board-specific overlays, which has the unfortunate side-effect that
forgetting to set this for a board results in a complete change in
behavior that is hard to notice.

This patch moves the whitelist to a single global file in the
laptop-mode-tools ebuild for all of ChromeOS. This has the advantage
that USB autosuspend is always off/whitelisted by default, and we only
have a single device database to maintain. Since many boards share
devices we at least have a chance to catch some of them in the whitelist
if we completely forget to add a board again. As we scale up, we might
be able to sync our PSO approved components list with this file to
support vendor-built devices by default, or even add a small set of
certified/tested external devices to improve power management.

This also removes an old patch concerning blacklisting of devices, since
that mode is no longer relevant.

BUG=chromium:264033
TEST=Plug an autosuspend-capable USB device into a stumpy, make sure
it's /sys/.../power/control node is set to 'on'.

Change-Id: Ic7c750e1c26c4429dacc299ee20fc46c870161d4
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63303
Reviewed-by: Todd Broch <tbroch@chromium.org>