build: Update scrollwheel acceleration curve.
This modifies the scrollwheel acceleration curve in
xf86-input-evdev. It increases the minimum scrolling step
and requires faster scrolling before acceleration really
begins to kick in.

The old patch claimed that the maximum distance to scroll
was 600 pixels, but as far as I can tell, it was actually
450. I'm maintaining the 450 limit in this change.

BUG=chromium:261626
TEST=manual: http://en.wikipedia.org/wiki/Cat

Change-Id: I139995594b25ace9d21acdd9b9cbdb537b8c5f5c
Reviewed-on: https://gerrit.chromium.org/gerrit/63401
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>