Move UI respawning logic to a separate upstart job
Sometimes a GPU hang will cause Chrome to be unable to start. In order to
mitigate these issues, we'd like to reboot the device when this happens.
A good indicator that we might be in this situation is the ui job exiting
with a certain status code. The difficulty is that, in order to take special
action on certain exit status codes, we need to take normal respawn handling
out of the hands of upstart as well.

This CL just installs the new upstart config files.

CQ-DEPEND=CL:62781
BUG=chromium:261784
TEST=build an image, check that /etc/init/ui-respawn.conf exists.

Change-Id: Ifedafeeb6c4ae6c65686fbe555f3596c3258dc05
Reviewed-on: https://gerrit.chromium.org/gerrit/63945
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>