chromeos-bootimage: Add option to build Early Firmware Selection image for ARM
Add a USE flag to enable building of an early-firmware-selection image for ARM
platforms where supported. This allows RO firmware to run from on-chip memory
so that RW firmare can initialize the SDRAM or adjust its settings before
starting up. This makes more the RW path updateable for both memory init and
the resume path.

BUG=chrome-os-partner:21115
TEST=manual

  . the generated efs image boots. More testing, including regular
    images, is required.

Change-Id: I12470f346341a83e841dfa4a9eb265ee4f5c1fa5
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64814