google-breakpad: build things in parallel and simplify
Rather than testing tc-is-cross-compiler to see if we're building for
the host, use the standard USE=cros_host flag.

Rather than create a copy of the source tree which we then build in
for the 32-bit version, build both versions out of tree.

By also using the multiprocessing eclass, we can run the two configure
scripts in parallel thus speeding up the overall build.

BUG=None
TEST=`sudo emerge google-breakpad` produces same files

Change-Id: I6c36fe8f05b50a4220c0c26718d21b74b33a0c45
Reviewed-on: https://gerrit.chromium.org/gerrit/65279
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>