P2P: Add unittests to the trybot step.
The p2p project has a considerable amount of unittest that are being
maintained and manually run on every commit to test the fixes.
Nevertheless, the trybot doesn't run those tests when a change to
chromeos-base/p2p is commited. This patch adds the function src_test
to the p2p ebuild in order to allow cros_run_init_tests to pick up
and run the unit tests.

This patch adds currently about 1 minute to the trybot running time
whenever the p2p project is modified.

BUG=None
TEST=Ran "cros_run_unit_tests --board=link -p chromeos-base/p2p". Test run and pass.

Change-Id: I4e6b32da1a4c865eda0e3b51a404ebb80892a97e
Reviewed-on: https://chromium-review.googlesource.com/65626
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>