google-breakpad: re-enable optimizations
The upstream bug does not appear to be an issue for us anymore.
We should at least try and see if things go south.

This also makes quite the difference for us at runtime: the time
needed to generate all symbols is cut in half (from ~3 min down
to ~1.5 min).

BUG=None
TEST=`sudo FEATURES=test emerge google-breakpad` passes 11 times in a row
TEST=`cbuildbot {arm,amd64,x86}-generic-full {x86-alex,lumpy}-release` pass

Change-Id: Id0b0424da011a9890e453e6390eebbd4ec2a3651
Reviewed-on: https://gerrit.chromium.org/gerrit/66458
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>