Build GCC libraries on a per-board basis.
Allow compiler libraries to be built on a per-board basis, rather than
on a per-architecture basis.
This CL contains the script to build the librarires.  THere are two
other related CLs for this as well:

CQ-DEPEND=CL:60289
CQ-DEPEND=CL:60303

BUG=chromium:197654
TEST=Built the comiler and libraries separately using these ebuilds and
the scripts in the other two CLs.  Successfully built and tested (with
the three CLs) on remote trybots for x86-alex-full, lumpy-full and
daisy-full.

Change-Id: Id9e6a166f79f074d839a05cd9278e82029a34ca0
Reviewed-on: https://chromium-review.googlesource.com/60350
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>