shorten autotest build-time; don't repackage packages all the time
The funcional part of this CL is in autotest.eclass. All the other
changes are just whitespace changes to force various ebuilds, which
inherit from the autotest eclass, to uprev.

Prior to this CL, there was some logic in the autotest eclass that would
switch over to using the autotest packager with the --all flag if it
detected more than some threshold of tests to package. However, as a
result of this, a small change in our build order resulted in the full
or nearly-full sysroot directory of tests being repackaged multiple
times, causing a major slowdown in the build time for autotest packages.

BUG=chromium:282663
TEST=Saw a preliminary speedup when I was cros_working on a few autotest
packages, with this CL in place. The true test will be seeing the
speedup once this change is merged and prebuilts are generated.

Change-Id: Iabcc54db7e0a10f2891cffe61393a8f92b67e563
Reviewed-on: https://chromium-review.googlesource.com/167939
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>