gcc: fix ASAN build for x86
Previously we pass -target TARGET to clang for cross compile.
for x86, the target is i686-pc-linux-gnu. When linking with asan,
it searchs for libclang_rt.asan-i686.a, which does not exist(we only
have libclang_rt.asan-i386.a). With this change, we do not use target
parameter, we use -m32 instead.

BUG=chromium:284940
TEST=build shill with asan clang on x86 and amd64
     build zlib with -clang-syntax on daisy

Change-Id: I53ef6c202ed4350d940d48ba1baeb9e23ed9241f
Reviewed-on: https://chromium-review.googlesource.com/168045
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>