gcc: use clang when '-fsanitize=address' in args.
We are using clang to generate code when ASAN is enabled.
If a package is ASAN enabled, then we are fine and this patch is
not needed. If a package is not ASAN enabled, it uses gcc instead of
clang to do the build, and the LDLFAGS is set to '-fsanitize=address'.
In this case, we have some problems.

BUG=None
TEST=cbuildbot chromiumos-sdk
Change-Id: Ifad4df7e332b797c66cefc8c43677a4265d16ba5
Reviewed-on: https://chromium-review.googlesource.com/66996
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>