Update the device_tree USE flag for benefit of beaglebone
This changes handling for the "device_tree" use flag in the
cros-kernel2 eclass when the flag is off.  This change has no effect
on existing boards, all of which build with "device_tree" on.

This is being made for the benefit of the new beaglebone board,
which requires it (for now, at any rate).

There are two functional changes:
 1) The 'dtb' make target is now built unconditionally, regardless
    of the use flag setting.
 2) With "device_tree" off, the .dtb files are installed under
    /boot/dts, which is where the beaglebone boot process expects to
    find its device tree file.

Building with "device_tree" off continues to build a uImage instead
of a zImage.

BUG=chromium:272337
TEST=build_packages

Change-Id: I6efd8ab8eabcc4f322af68889670e10df7154245
Reviewed-on: https://chromium-review.googlesource.com/168685
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>