autotest.eclass: fix a semantic error
This CL fixes a long-term syntax error in autotest.eclass, by
eliminating the obsolete if test that it is part of. It also
uprevs a bunch of autotest-eclass inheriting ebuilds, and makes it
possible to specify values of AUTOTEST_PROFILERS_LIST other than * or
null.

BUG=chromium:286109
TEST=trybots

Change-Id: I2f957a940b0bc0d5204c291262c2ddc2e25c73a5
Reviewed-on: https://chromium-review.googlesource.com/168819
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>