Update libcurl from 7.23.1 to 7.31.0.
This patch updates libcurl to the latest stable ebuild available on
Gentoo. The version 7.32.0 doesn't introduce any security fix over
7.31.0 and is marked as unstable, so this CL upgrades to 7.31.0.
This patch also disables the static-libs use flag for curl, since
the required openssl dependency is not being compiled with
static-libs.

This update removes the following patches because they are merged
upstream (released in 7.24.0):
 * 0001-CURLOPT_DNS_SERVERS-set-name-servers-if-possible.patch
 * 0001-CURLOPT_DNS_SERVERS-set-name-servers-if-possible-fix.patch
 * 0001-Do-not-try-to-do-DNS-name-resolution-on-interface-na.patch
 * 0001-multi-handle-timeouts-on-DNS-servers-by-checking-for.patch
 * 0001-multi-interface-only-use-non-NULL-function-pointer.patch

The following patch removes the @LDFLAGS@ argument on the --libs)
case on the curl-config.in file; but this argument is not present
anymore on the 7.31.0 version. Removing it from the list:
 * curl-7.20.0-strip-ldflags.patch

A new patch is included from upstream together with the ebuild and
another patch is updated:
 * curl-7.30.0-prefix.patch (updated patch)
 * curl-fix-gnutls-nettle.patch (new patch)

Finally, a patch included in 7.32.0 is added to fix the bug described
in http://sourceforge.net/p/curl/bugs/1249/ that is being hit in our
test infrastructure:
 * 0001-ossl_recv-check-for-an-OpenSSL-error-don-t-assume.patch

BUG=chromium:237562
TEST=Trybot passes. Manual check that update_engine updates to next build from my desktop.

Change-Id: I9cb5c3fd0f21e3201974be8258eba49dd952a15d
Reviewed-on: https://chromium-review.googlesource.com/169435
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>