Install new env vars header
We'd like for the session manager to be able to communicate with
breakpad, in order to inject a message into browser crash reports
in the event that the browser needed to be killed for some reason.
We're passing the number of the FD as an env var to remain consistent
with the mechanism used in the linux suid sandbox code.

This change installs a header that defines the name of this environment
variable, and which can be used to define the names of other such shared
vars in the future.

BUG=chromium:233833
TEST=emerge system_api

CQ-DEPEND=I322e3fe0888509e8bbd9c15e0c9137eb4e63e6e0

Change-Id: If6ca9a8d0ac73785964de351cc31572c6b88d005
Reviewed-on: https://chromium-review.googlesource.com/169439
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>