Add a chromeos-activate-date ebuild and enable it through vpd
In many cases we will want to record the activation date for a device,
some devices do this already in their board specific overlays, however
as this is a common request we should do it in one place in a standardized
way across devices. We still enable custom activate_date scripts to be
put in by custom board overlays if needed with a virtual override.

We rdepend this on the vpd ebuild as any system using vpd should save the
activate date in vpd.

BUG=chromium:289118
TEST=Run on a DUT and verify the ActivateDate RW_VPD entry gets set.
CQ-DEPEND=CL:169819
CQ-DEPEND=CL:169860
CQ-DEPEND=CL:*144872
CQ-DEPEND=CL:*144979
CQ-DEPEND=CL:*144985

Change-Id: Iba5508f00bc1be59807350e2770553fe177f37ba
Reviewed-on: https://chromium-review.googlesource.com/169650
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>