bootimage: Prevent loading the env on new U-Boot (vboot)
On the old U-Boot (chromeos-v2011.12), the loading of the environment
was disabled by default and was enabled with the "load_env" node.  On
new U-Boot the loading of the environment is enabled by default and is
disabled with "load-environment".

We only need to support the new U-Boot (old versions are built on
branches), so modify the source to use the new node name.

BUG=chrome-os-partner:22473
TEST=Boot U-Boot and don't see environment CRC error; can't change env
TEST=Boot nv-U-Boot and see that env is loaded

Change-Id: I85245f1fa0047e6e97c72f968c0d52245b70ee65
Reviewed-on: https://chromium-review.googlesource.com/169649
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>