CHROMIUM: tlsdate update to refactor
Change to support the new configuration arguments
and use of NDEBUG.

Also switch to proxystate as the DBus owner so that we
can merge the platform changes without a specific
Chrome permissions dependency. After, we can change
all the places to tlsdated or tlsdate-dbus and remove
the proxystate user.

BUG=chromium:293695
TEST=builds and installs
CQ-DEPEND=CL:169131,CL:170566,CL:170650
Depends-On: I0d24951182a768532d6d04eea2224ac70cf04052
Depends-On: Ifd0cb6feaae4246c91a3c06738cf8a443706ed6c
Change-Id: I748648918b7f37984590e7d0c8081f3981d2b725
Reviewed-on: https://chromium-review.googlesource.com/170601
Reviewed-by: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>