Revert "libxml2: move to portage-stable"
This reverts commit fcd648e544e176b4a4078c98691eb1d84b9713a0.

The cros-sdk is still using this older version and w/out it,
the build breaks:
http://build.chromium.org/p/chromiumos/builders/chromiumos%20sdk/builds/4030/steps/InitSDK/logs/stdio

This is due to the new libxml2 not supporting USE=python, but
some of the build tools (notably, app-text/gnome-doc-utils)
want that enabled.

So re-enable this old version for the sdk.  Our boards still 
use the new version which is what we really care about security
wise.

Change-Id: I83a5d47b16d5e31438a1d07decb6c5f2b9649e6a
Reviewed-on: https://chromium-review.googlesource.com/169604
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>