Revert "autotest-deps-glbench: enable asan check."
It turns out that the autotest-deps-glbench ignores user
CFLAGS at the compile time.
So the clang-syntax checking is not enabled and there are
several clang-syntax checking errors in this package.
New clang with ASAN will complain if '-fsanitize=address'
does not appear both at compile and link time.
So revert this first and try to roll to new clang. After that
we can fix errors in this package later.

This reverts commit e4706af63012a9993800878d88136dc1f042aa0f.

BUG=None
TEST=None

Change-Id: Ib36876df3ad88e5d5d6437312d0ad1cb79d4935e
Reviewed-on: https://chromium-review.googlesource.com/170516
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>