cros-kernel: mung vdso names to help out debuggers
Rather than parse the ELF and mung the dumped symbol file (see CL:170672),
install the vdso with names & symlinks as other tools expect.

For breakpad, it wants things named "linux-gate.so".  For gdb (and things
parsing the ELF SONAMEs), we use "linux-gate.so.1" and "linux-vdso.so.1".

It does mean slightly uglier file names, but now we don't have to modify
the existing tool infrastructure.

BUG=chromium:220119
TEST=`emerge-x86-alex chromeos-kernel` installs vdso files as linux-gate.so

Change-Id: I9f1e10fc3d790757d1f1b2cd8b80e7ecf219458b
Reviewed-on: https://chromium-review.googlesource.com/170725
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>