chromeos-test: Add dev-python/dpkt to test images.
The dpkt python module is already included in autotest-deps, because
it is used on the host during compile and unittest of the 'lansim'
autotest dep. It also needs to be installed on the test images to
be available for tests during runtime, since dpkt is not being
packaged as part of lansim.

BUG=chromium:288385
TEST=build a test image and verify dpkt is present on /usr/local python path.

Change-Id: I75a2cb447daf1c1eb992e22ab6c2587eef7f0eaa
Reviewed-on: https://chromium-review.googlesource.com/171005
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>