udev: set group read permission for tty1 and tty2
X Server needs to open /dev/tty1 RW on startup. Right now TTYs are
set to group write only. When X server will be started as xorg user
which belongs to tty group, we need tty1 to be group RW. We also
set tty2 because in some test scenarios X server can run on it.

BUG=chromium:182119
TEST=login to device, ls -al /dev/tty1 /dev/tty2

Change-Id: Ifffd6159a6fb15c11f21c752f66eb86bb399f7d8
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/171691
Reviewed-by: Mike Frysinger <vapier@chromium.org>