Turn off the merge-sync feature, which slows down emerge dramatically.
The merge-sync feature slows down emerge quite a bit, because it forces
portage to sync to disk after every merge. In new versions of emerge,
this is on by default, so we should disable it.

Thanks to vapier@ for help setting up a bisect script to find this
issue.

BUG=chromium:218085, chromium:268684
TEST=Run locally with new portage, notice faster emerge
TEST=old portage still works

Change-Id: Icda5d5ac8e97fc9b123655b1416ed976679f3cad
Reviewed-on: https://chromium-review.googlesource.com/171773
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>