Re-land "portage: upgrade to current upstream stable"
This upgrades us to Portage 2.1.11.50, which supports
EAPI=5. This requires a two-phase update, so I've left
the old version of Portage in place.

The previous incarnation of this CL slowed down merges
significantly because Portage 2.1.11.50 turns on
merge-sync by default. Disabling that fixed the issue,
so we can land the upgrade again.

This CL is identical to CL:42993, except that it has
been rebased on top of CL:171773.

BUG=chromium:218085
TEST=Major release, full, paladin, and sdk trybots work.
TEST=Local builds are just as fast as before.

Change-Id: I7f74aef52044d373f54f79117867df622d0f07d4
Reviewed-on: https://chromium-review.googlesource.com/171794
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>