Add new chromeos-factoryinstall-bsp package
Currently chromeos-bsp is shared by both base image and install shim.
This results in increasingly number of files that are only used by the
factory showing up in base image.

To solve this, let's add a new chromeos-factoryinstall-bsp package
that only chromeos-installshim depends on. Files that we don't want to
put into base image go into this package.

BUG=chrome-os-partner:19771
TEST=emerge-daisy_spring chromeos-installshim. See chromeos-factoryinstall-bsp
merged.

Change-Id: I0c657dec40a3d95b1c20aac2e059daec7bf0fbed
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172410