dpkt: Add a fix to import ip.
When dpkt.ip is imported on a python script, it will try to import
all the IP protocols implemented on dpkt, but, if a package with the
same name as the desired protocol is found on the path it could pick
that one and fail instead of using the one provided with dpkt.

This patch adds a fix that forces the import to search only on the
same directory as dpkt.ip.
Patch submitted upstream.

BUG=chromium:271847
TEST="emerge-daisy_spring dpkt" shows the patch applied:
 * Applying dpkt-1.8-fix-ip-import.patch

Change-Id: I390fff0b460c9fb3ac8b759bf28a40320d0e44ed
Reviewed-on: https://chromium-review.googlesource.com/172650
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>