Chromeos-chrome don't copy broken symlinks
install_test_resources copies over a list of files/folders from
chrome into a test_src directory to be packaged as an autotest
dep. This can result in broken symlinks if a symlink is copied
but the file it points to is not. This change simply removes
broken symlinks from the collection of copied files.

Also adds tools/perf/run_measurement into the telemetry dep.

BUG=chromium:305873, chromium:305754
TEST=locally built and the broken symlinks no longer exist. Did a
trybot run to make sure nothing broke as well.

Change-Id: I4b5dc3665d20418f50e79ded28484de8d3e08089
Reviewed-on: https://chromium-review.googlesource.com/172695
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>