gcc: include $PVR in version output
So that tools (like GOMA) can detect when the toolchain has changed in
a meaningful way, include the portage PVR in the version string.  This
might mean we sometimes bump when we don't need to, but much more often
than not, it does matter.  Let's not waste developer time figuring out
if they need to actually manually bump it.

BUG=chromium:306101
TEST=`sudo emerge cross-i686-pc-linux-gnu/gcc && i686-pc-linux-gnu-gcc --version` shows $PVR

Change-Id: I28d24ef03b66fb8e6b5a0697c7d15de7a7e776a0
Reviewed-on: https://chromium-review.googlesource.com/172841
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>