portage: re-order install hooks execution
We want the install hooks to run before stripping so we can do our own
QA checks on them.  Bump the order so that happens.

In the future, we'll want to re-architect this code entirely, but the
qa checking code in portage will need a rewrite.  See BUG for more
details.

BUG=chromium:217578
TEST=`emerge-x86-alex power_manager` shows flag stuff works and things still get stripped
TEST=`cbuildbot chromiumos-sdk` works

Change-Id: Ic3ac31e194c7eb1a4973ba9f5ceaa14f5ea71530
Reviewed-on: https://chromium-review.googlesource.com/175049
Reviewed-by: Marc MERLIN <merlin@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>