chrome: stop double-stripping of autotest files
The current logic strips the autotest code and then turns around and
strips the whole $D dir implicitly (when `prepstrip` runs at the end).

This was added when deploy_chrome was first added and doing the strip,
but later we disabled stripping in deploy_chrome w/out also disabling
this new bit of code.

BUG=chromium:305263
TEST=`emerge-x86-alex chromeos-chrome` works and doesn't double strip

Change-Id: I07f192fbebdb007bc72a0caefd01f6241ea81e83
Reviewed-on: https://chromium-review.googlesource.com/173615
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>