Provide a facility to build coreboot with reduced console output
Default coreboot console settings result in tons of serial port
output, which makes booting much longer while not being very
interesting if it is not coreboot which is being debugged.

With the current approach to storing board configs including only
non-default settings, it becomes easy to modify the stored config to
change the default console output level.

This change is adding a use flag (quiet-cb) which reduces the default
coreboot output level to 3.

BUG=none
TEST=manual

   . build and run Bayley Bay coreboot with and without USE=quiet-cb,
     observe different amount of console output generated during
     startup.

Change-Id: I88e61cbb027e02b9e97d10e14e7e3a6a7f97cafe
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/175059
Reviewed-by: Stefan Reinauer <reinauer@google.com>