Select xorg configs between Spring at Daisy
Now that the new Lucas VE1 is going to be shipping with Atmel we need to
add in the Xorg configs for it.  However it's a slightly different pad
than Spring so we can't just use the same one.  Due to the way the
device matching works, we can't automatically tell them apart in X so
this CL splits them into two files.  The normal daisy.conf contains the
configs for Cypress and Atmel versions of Lucas where spring.conf
contains only the config for Spring

This chooses between the two configs and makes sure they end up with the
right configs.

BUG=chrome-os-partner:23912
TEST=I built ChromeOS with these patches for both Spring and Lucas and
the correct configs were loaded.  The touchpads seem to work fine with
these configs
CQ-DEPEND=I168b6261f758ca74de383cb33c61d73b736e55cc

Change-Id: Iba21ca42226aa727c249a50df0a740333164b6df
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/176252
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>