neard-configs: Create and install /etc/neard/main.conf.
Added a new configuration file, which puts neard into "ConstantPoll"
mode. The default behavior of the NFC device is to stop polling for new devices
once a connection to a tag or peer device is established. This feature tells
neard to automatically restart the poll-loop if it loses the connection with
a device or a tag. This way the D-Bus client won't have to explicitly call
org.neard.Adapter.StartPollLoop to get NFC running again.

BUG=chromium:317878
CQ-DEPEND=I9c99383e1885edd04ec68a02687036d56043acd1
TEST=Verify using /usr/local/autotest/cros/nfc/console.py.

Change-Id: I3e323c486caa396bf3af7a8e8ea9017f23f5b050
Reviewed-on: https://chromium-review.googlesource.com/176460
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Arman Uguray <armansito@chromium.org>