Introduce virtual/implicit-system
In the interest of capturing all the build/runtime dependencies in
CrOS that should 'go without saying', we're introducing a virtual
called 'implicit-system'. That is the system-layer stuff that you
can just go ahead and assume is there.

While portage supports a notion of 'sets', we chose a virtual package
because sets:
 - lack any sort of visibility in the CrOS community,
 - don't allow anything fun/dynamic (like leverage USE flags),
 - can be replicated via metapackages, and
 - aren't used in CrOS today, so it's a completely new (and different) concept, and unlike anything else we're doing.

BUG=None
TEST=do a full build and run suite:smoke

Change-Id: I6a641867748789b503544ac5e3d59315998b6c63
Reviewed-on: https://chromium-review.googlesource.com/177055
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>