Add a package to add .ssh/authorized_keys to /root
This package installs authorized_keys to /root/.ssh.
Previously the authorized_keys is installed by a script
mod_image_for_test_scripts/350addTestAuthKeys
However, that script won't get called during the
building process of a beaglebone image. As part of the change
that is needed to solve the problem, this cl moves
the installation of authroized_key to a new package called
chromeos-base/chromeos-test-testauthkeys/.
With this change, any other packages that needs to
install authorized_keys can add a dependency of this package.

BUG=chromium:306619
TEST=1)Ran emerge-beaglebone chromeos-base/chromeos-test-testauthkeys
Confirmed the file is installed at
/build/beaglebone/root/.ssh/authorized_keys.
2) Ran emerge-lumpy chromeos-base/chromeos-test-testauthkeys
Confirmed the file is installed at
/build/lumpy/root/.ssh/authorized_keys.
3) Added a dependency to this package in
chromeos-base/chromeos-bsp-beaglebone,
built a beaglebone image, confirmed the beaglebone image
had the correct ssh keys. I could ssh to the beaglebone as root without
password.

Change-Id: I8a7f9f365b09d3086f5e4ce5d9b987053c8e4a05
Reviewed-on: https://chromium-review.googlesource.com/177426
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>